Expanded Transform & Camera
Tried to implement Camera movement but had to basically rewrite
everything to get it to work. Transforms now have local directions which
allows for simply forward, backward, left, etc. movement. The Camera
implements this sytem to move around the world, as well. The Game Input
had a bug where when initialized, it would consider all keys to be down,
that is now fixed. Lastly, got rid of the old useless shaders.