Replaced Line Operations With Post Processing
Replaced the previously added line operations in the Shader class with
post processing. This is due to the line operation performance boost
being very minute and only adding a lot of overhead, but also makes the
Shader uniform detection a lot more powerful as uniform declarations can
span multiple lines.