Updated Transform's Transform and Set Translation
Updated the Transform's TransformPoint and InverseTransformPoint methods
to rely on the rotation and scale of the object rather than just the
position, and the Set Translation method to account for the new changes
of above methods. Also fixed some errors in the Matrix multiplication
algorithms and in the Game Object's unparent procedure.