Rewrote the Test Game
Cleaned out all of the debug code from the Test Game and rewrote it to
be based off of a new Scene. Fixed a bug in the Scene class where
children lights would not be properly added to the active lights list.
Also removed resource files from GitIgnore since the Test Game requires
them to run.