Fixed Blinn-Phong and Specular GBuffer
Removed an unnecessary negative from the Blinn-Phong specular model to
fix the specular lighting for all lighting types. Also changed the
Specular Buffer's texture type to RGBA16F to allow for a wider array of
values in the alpha component of the texture, which is where we store
the material shininess.