Fixed Blinn Phong Specular + Changed Scene
Fixed an issue where the Specular Reflection would be calculated even if
the light was coming from underneath a surface by implementing the
diffuse factor into the equation. Also changed the scene a little bit to
test around the Point Light and Spot Light components on an Input
Transformable Suzanne.