Implemented Normal Mapping
Implemented and tested normal mapping. Changed the normals calculation
method in the Geometry Fragment Shader to automatically select the bump
map normal or the interpolated vertex shader normal depending on the
Material properties. Also made some slight adjustments to the Model and
Texture classes, to allow for the creation of non-white 1x1 pixel
textures, which is utilized by the normal calculation algorithm in the
Geometry Shader.