Added Helper Methods to Model
Added two helper methods for fetching material properties from ASSIMP to
eliminate some of the code repetition and to check if the material
property was successfully fetched. Also fixed an error in the reflection
calculation of the geometry fragment shader where the reflection was not
being clamped, resulting in a reflection appearing when it shouldn't.