Implemented Albedo and Diffuse Map for Material
Replaced the old diffuse texture/color with albedo, where albedo refers
to the base texture/color of the material. The new diffuse now refers to
the way diffuse light affects the material. This aligns the engine more
into the direction of PBR and should also make it easier to implement
ambient and emissive material maps.