Added Specular Lighting Material Data to AAD Buffer
Added the material's specular color and shininess data to the AAD
Buffer, creating a new Material Buffer. This allows all of the (current)
material data to be stored in a single buffer. Also changed the packing
algorithm to rely on floats again since sampling a texture returns a
highp floating point vector rather than a uint.