Added Clean Up Code to All Renders in Deferred
Added some small bits of clean up code to all of the render methods of
the Deferred Renderer. Now if any render function changed an internal
OpenGL function, it undoes the change at the end of its execution. This
helps make all of the render functions independent from each other.