Added Equality Operators for Textures & Cubemaps
Added (in)equality operator overloading for the Texture2D and Cubemap
classes. Also removed the width and height fields from the Texture2D
class since those are based off of the data stored stored in OpenGL
rather than the data stored in a specified instance of the Texture2D
class.