Made Transform a Default Component
Made Transform a default component so that any new Game Object created
automatically gets a Transform component attached to it. Also added a
GameComponentPermanentException which will be thrown when attemping to
remove a permanent Game Component. Transforms are also now permanent and
cannot be removed once added to a Game Object.