Refactored MeshRenderer Component
Refactored the MeshRenderer component. The MeshRenderer now stores a
list of the Meshes and Materials which it should render, which allows
one component attached to an object to render multiple meshes. This
replaces the old system which required multiple MeshRenderers to be
attached to the same object to render multiple meshes.