Refactored Model and GameObjectFactory
Refactored the Model and GameObjectFactory classes. The Model class now
has XML comments and it only stores the Game Object from which clones
are made, it is no longer necessary for the Model class to retain
ownership of all of the Model's resources, making it a lot cleaner and
safer.