Changed GameObject's GetComponents Method
Changed the GameObject's GetComponents method to return a vector of
downcast references rather than a vector of raw pointers. This should
help make the method easier to work with since it is no longer necessary
to manually downcast each element in the vector. Also fixed a misc.
iterator issue in the EventHandler class.